chore(rename): page folder routing convention
