+++ src/pages/User/sign-up/hooks/useRegisterForm.ts
@@ -4,7 +4,7 @@ import { useForm } from 'react-hook-form';
 
 import { PATH_ROOT } from '@/const/path';
 import { REGISTER_FORM_KEY } from '@/const/register-form';
-import useRegisterResolver from '@/pages/User/SignUp/hooks/useRegisterResolver';
+import useRegisterResolver from '@/pages/user/sign-up/hooks/useRegisterResolver';
 import { useServiceRegister } from '@/services/auth/services';
 
 export type TRegisterFormFields = {
