+++ src/pages/about-us/components/SwipeNavigation/index.tsx
